fix utf-8 downgrade fatal error
When Mojo::DOM parses the page, it unescapes HTML entities. This causes
high-byte characters in the string. Then when we stringify it, the
entities are not re-escaped, leaving those high-byte characters. When
Mojo::IOLoop::Stream tries to utf8::downgrade with those high-byte
characters, it throws a fatal error.

Now we just parse HTML with regex, which is much better.