rewrite poller to prevent tying up workers
Since Mojolicious only uses PSGI in a blocking way, every long-polling
request ties up a worker for the duration. I could not figure out a way
to detect if the client was still connected and stop the poller on the
server side, so just navigating through the site quickly caused all the
workers to be used up...

This new mechanism does short polling at quick intervals. Since I hook
around the dispatch, these requests do not show up in the logs at all,
so I can make dozens of requests without annoying spam in the logs.